# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin-Panel-Api",
        "description": "Laravel Api",
        "contact": {
            "name": "Hamed",
            "email": "h.akbarzadh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/store/store/{product}": {
            "post": {
                "tags": [
                    "Store"
                ],
                "summary": "Store product_marketable in DB",
                "description": "Store product_marketable in DB",
                "operationId": "storeStore",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "marketable_number",
                                    "receiver",
                                    "delivery",
                                    "description"
                                ],
                                "properties": {
                                    "receiver": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Alex"
                                    },
                                    "delivery": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "David"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    },
                                    "marketable_number": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/store/update/{product}": {
            "put": {
                "tags": [
                    "Store"
                ],
                "summary": "Update Store in DB",
                "description": "Update Store in DB",
                "operationId": "storeUpdate",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "marketable_number",
                                    "receiver",
                                    "delivery",
                                    "description"
                                ],
                                "properties": {
                                    "receiver": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Alex"
                                    },
                                    "delivery": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "David"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    },
                                    "marketable_number": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/store/get-store/{number}": {
            "get": {
                "tags": [
                    "Store"
                ],
                "summary": "Get Store Detail",
                "description": "Get Store Detail",
                "operationId": "storeShow",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Id of brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}